1. WeatherInfo 클래스 분리 (책임 분리)
